- Moved code related to generating command-line args into ConfigRunner.
